# [hw 10] Program optimization