--- a/.github/workflows/tests.yml
+++ b/.github/workflows/tests.yml
@@ -27,7 +27,7 @@ jobs:
       - name: Linters
         uses: golangci/golangci-lint-action@v3
         with:
-          version: v1.50.1
+          version: 1.54.2
           working-directory: ${{ env.BRANCH }}
 
   tests: