--- a/.github/workflows/tests.yml
+++ b/.github/workflows/tests.yml
@@ -27,7 +27,7 @@ jobs:
       - name: Linters
         uses: golangci/golangci-lint-action@v3
         with:
-          version: 1.54.2
+          version: v1.54.2
           working-directory: ${{ env.BRANCH }}
 
   tests: