--- a/.gitignore
+++ b/.gitignore
@@ -1,5 +1,7 @@
 ### Go template
 # Binaries for programs and plugins
+bin/
+
 *.exe
 *.exe~
 *.dll