--- a/hw10_program_optimization/Makefile
+++ b/hw10_program_optimization/Makefile
@@ -1,7 +1,7 @@
 LOCAL_BIN:=$(CURDIR)/bin
 GOLANGCI_TAG:=1.54.2
 MEMORY_PROFILE_FILE=memory.out
-CPU_PROFILE_FILE=memory.out
+CPU_PROFILE_FILE=cpu.out
 
 .PHONY: .bin-deps
 .bin-deps: