--- a/hw10_program_optimization/Makefile
+++ b/hw10_program_optimization/Makefile
@@ -31,7 +31,7 @@ easyjson:
 
 .PHONY: bench
 bench:
-	go test -v -count=1 -timeout=30s -tags bench . -memprofile=$(MEMORY_PROFILE_FILE) -cpuprofile $(CPU_PROFILE_FILE)
+	go test -v  -tags bench -bench . -benchmem -run BenchmarkGetDomainStat -memprofile=$(MEMORY_PROFILE_FILE) -cpuprofile $(CPU_PROFILE_FILE)
 
 .PHONY: clear_bench
 clear_bench: