--- a/hw10_program_optimization/progres.txt
+++ b/hw10_program_optimization/progres.txt
@@ -1,4 +1,4 @@
-Start:
+# Start:
 
 ```
 goos: linux
@@ -11,7 +11,7 @@ PASS
 ok      github.com/hound672/otus-hw/hw10_program_optimization   7.447s
 ```
 
-With easyjson:
+# With easyjson:
 
 ```
 go test -v  -tags bench -bench . -benchmem -run BenchmarkGetDomainStat -memprofile=memory.out -cpuprofile cpu.out
@@ -25,7 +25,7 @@ PASS
 ok      github.com/hound672/otus-hw/hw10_program_optimization   5.738s
 ```
 
-Compile regexp:
+# Compile regexp:
 
 ```
 go test -v  -tags bench -bench . -benchmem -run BenchmarkGetDomainStat -memprofile=memory.out -cpuprofile cpu.out
@@ -40,3 +40,17 @@ ok      github.com/hound672/otus-hw/hw10_program_optimization   3.122s
 ```
 
 ### After compile regexp bench passed by cpu, but not memory
+
+# Use scanner
+
+```
+go test -v  -tags bench -bench . -benchmem -run BenchmarkGetDomainStat -memprofile=memory.out -cpuprofile cpu.out
+goos: linux
+goarch: amd64
+pkg: github.com/hound672/otus-hw/hw10_program_optimization
+cpu: AMD Ryzen 7 PRO 3700 8-Core Processor
+BenchmarkGetDomainStat
+BenchmarkGetDomainStat-16            165           6580035 ns/op           93658 B/op       4536 allocs/op
+PASS
+ok      github.com/hound672/otus-hw/hw10_program_optimization   2.822s
+```