--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -8,6 +8,7 @@ import (
 	"strings"
 )
 
+//easyjson:json
 type User struct {
 	ID       int
 	Name     string