--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -49,12 +49,13 @@ func getUsers(r io.Reader) (result users, err error) {
 
 func countDomains(u users, domain string) (DomainStat, error) {
 	result := make(DomainStat)
+	reg, err := regexp.Compile("\\." + domain)
+	if err != nil {
+		return nil, fmt.Errorf("regexp.Compile: %w", err)
+	}
 
 	for _, user := range u {
-		matched, err := regexp.Match("\\."+domain, []byte(user.Email))
-		if err != nil {
-			return nil, err
-		}
+		matched := reg.Match([]byte(user.Email))
 
 		if matched {
 			num := result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])]