--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -15,47 +15,37 @@ type User struct {
 type DomainStat map[string]int
 
 func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
-	u, err := getUsers(r)
-	if err != nil {
-		return nil, fmt.Errorf("get users error: %w", err)
-	}
-	return countDomains(u, domain)
-}
-
-type users []User
-
-func getUsers(r io.Reader) (users, error) {
 	s := bufio.NewScanner(r)
-	result := make(users, 0)
+	stat := DomainStat{}
 
 	for s.Scan() {
-		var user User
+		user := &User{}
 		if err := user.UnmarshalJSON(s.Bytes()); err != nil {
 			return nil, fmt.Errorf("user.UnmarshalJSON: %w", err)
 		}
-		result = append(result, user)
+
+		dom := getDomainFromEmail(user.Email, domain)
+		if dom == "" {
+			continue
+		}
+
+		num := stat[dom]
+		num++
+		stat[dom] = num
 	}
 
-	return result, nil
+	return stat, nil
 }
 
-func countDomains(u users, domain string) (DomainStat, error) {
-	result := make(DomainStat)
-
-	for _, user := range u {
-		if !strings.HasSuffix(user.Email, "."+domain) {
-			continue
-		}
-
-		sp := strings.SplitN(user.Email, "@", 2)
-		if len(sp) != 2 {
-			continue
-		}
-
-		dom := strings.ToLower(sp[1])
-		num := result[dom]
-		num++
-		result[dom] = num
+func getDomainFromEmail(email, domain string) string {
+	if !strings.HasSuffix(email, "."+domain) {
+		return ""
 	}
-	return result, nil
+
+	sp := strings.SplitN(email, "@", 2)
+	if len(sp) != 2 {
+		return ""
+	}
+
+	return strings.ToLower(sp[1])
 }