--- a/hw10_program_optimization/stats_optimization_test.go
+++ b/hw10_program_optimization/stats_optimization_test.go
@@ -46,7 +46,7 @@ func TestGetDomainStat_Time_And_Memory(t *testing.T) {
 	t.Logf("time used: %s / %s", result.T, timeLimit)
 	t.Logf("memory used: %dMb / %dMb", mem/mb, memoryLimit/mb)
 
-	// require.Less(t, int64(result.T), int64(timeLimit), "the program is too slow")
+	require.Less(t, int64(result.T), int64(timeLimit), "the program is too slow")
 	require.Less(t, mem, memoryLimit, "the program is too greedy")
 }