## Домашнее задание №10 «Оптимизация программы»

### Чек-лист студента ([Что это?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0))
- [x] Я перечитал задание после решения. [Зачем перечитывать задание?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0#user-content-%D0%97%D0%B0%D1%87%D0%B5%D0%BC-%D0%BF%D0%B5%D1%80%D0%B5%D1%87%D0%B8%D1%82%D1%8B%D0%B2%D0%B0%D1%82%D1%8C-%D0%B7%D0%B0%D0%B4%D0%B0%D0%BD%D0%B8%D0%B5)
- [x] Я запустил `go mod tidy`.
- [x] Я удалил `.sync` файл. [Зачем его удалять?](https://github.com/OtusGolang/home_work/wiki/%5B%D0%A1%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0%D0%BC%5D-%D0%9F%D1%80%D0%BE%D1%86%D0%B5%D1%81%D1%81-%D1%81%D0%B4%D0%B0%D1%87%D0%B8-%D0%94%D0%97#user-content-%D0%92%D0%B0%D1%80%D0%B8%D0%B0%D0%BD%D1%82-2)
- [x] Я написал бенчмарк и приложил результаты benchstat.

### Критерии оценки
- [ ] Пайплайн зелёный и нет попытки «обмануть» систему - 4 балла
- [ ] Добавлены юнит-тесты - до 3 баллов
- [ ] Понятность и чистота кода - до 3 баллов

#### Зачёт от 7 баллов

---

# Start:

```
goos: linux
goarch: amd64
pkg: github.com/hound672/otus-hw/hw10_program_optimization
cpu: AMD Ryzen 7 PRO 3700 8-Core Processor
BenchmarkGetDomainStat
BenchmarkGetDomainStat-16            324           3190594 ns/op          978681 B/op       8789 allocs/op
PASS
ok      github.com/hound672/otus-hw/hw10_program_optimization   7.447s
```

# With easyjson:

```
go test -v  -tags bench -bench . -benchmem -run BenchmarkGetDomainStat -memprofile=memory.out -cpuprofile cpu.out
goos: linux
goarch: amd64
pkg: github.com/hound672/otus-hw/hw10_program_optimization
cpu: AMD Ryzen 7 PRO 3700 8-Core Processor
BenchmarkGetDomainStat
BenchmarkGetDomainStat-16            458           2435046 ns/op          634237 B/op       5561 allocs/op
PASS
ok      github.com/hound672/otus-hw/hw10_program_optimization   5.738s
```

# Compile regexp:

```
go test -v  -tags bench -bench . -benchmem -run BenchmarkGetDomainStat -memprofile=memory.out -cpuprofile cpu.out
goos: linux
goarch: amd64
pkg: github.com/hound672/otus-hw/hw10_program_optimization
cpu: AMD Ryzen 7 PRO 3700 8-Core Processor
BenchmarkGetDomainStat
BenchmarkGetDomainStat-16            548           2103875 ns/op          281954 B/op       1546 allocs/op
PASS
ok      github.com/hound672/otus-hw/hw10_program_optimization   3.122s
```

### After compile regexp bench passed by cpu, but not memory

# Use scanner

```
go test -v  -tags bench -bench . -benchmem -run BenchmarkGetDomainStat -memprofile=memory.out -cpuprofile cpu.out
goos: linux
goarch: amd64
pkg: github.com/hound672/otus-hw/hw10_program_optimization
cpu: AMD Ryzen 7 PRO 3700 8-Core Processor
BenchmarkGetDomainStat
BenchmarkGetDomainStat-16            165           6580035 ns/op           93658 B/op       4536 allocs/op
PASS
ok      github.com/hound672/otus-hw/hw10_program_optimization   2.822s
```

# Remove not used fields from struct

```
go test -v  -tags bench -bench . -benchmem -run BenchmarkGetDomainStat -memprofile=memory.out -cpuprofile cpu.out
goos: linux
goarch: amd64
pkg: github.com/hound672/otus-hw/hw10_program_optimization
cpu: AMD Ryzen 7 PRO 3700 8-Core Processor
BenchmarkGetDomainStat
BenchmarkGetDomainStat-16            612           1651633 ns/op           16118 B/op        419 allocs/op
PASS
ok      github.com/hound672/otus-hw/hw10_program_optimization   2.217s
```

# Use dynamic slice

```
go test -v  -tags bench -bench . -benchmem -run BenchmarkGetDomainStat -memprofile=memory.out -cpuprofile cpu.out
goos: linux
goarch: amd64
pkg: github.com/hound672/otus-hw/hw10_program_optimization
cpu: AMD Ryzen 7 PRO 3700 8-Core Processor
BenchmarkGetDomainStat
BenchmarkGetDomainStat-16         291168              3978 ns/op            5557 B/op         19 allocs/op
PASS
ok      github.com/hound672/otus-hw/hw10_program_optimization   3.827s
```

# Remove regexp

```
go test -v  -tags bench -bench . -benchmem -run BenchmarkGetDomainStat -memprofile=memory.out -cpuprofile cpu.out
goos: linux
goarch: amd64
pkg: github.com/hound672/otus-hw/hw10_program_optimization
cpu: AMD Ryzen 7 PRO 3700 8-Core Processor
BenchmarkGetDomainStat
BenchmarkGetDomainStat-16         802689              1458 ns/op            4159 B/op          2 allocs/op
PASS
ok      github.com/hound672/otus-hw/hw10_program_optimization   3.519s
```

# Remove slice

```
go test -v  -tags bench -bench . -benchmem -run BenchmarkGetDomainStat -memprofile=memory.out -cpuprofile cpu.out
goos: linux
goarch: amd64
pkg: github.com/hound672/otus-hw/hw10_program_optimization
cpu: AMD Ryzen 7 PRO 3700 8-Core Processor
BenchmarkGetDomainStat
BenchmarkGetDomainStat-16         834274              1420 ns/op            4147 B/op          2 allocs/op
PASS
ok      github.com/hound672/otus-hw/hw10_program_optimization   3.722s
```